feat: 습관 관리 모달 로직을 커스텀 훅으로 분리
